# lib/api-zod/src/generated/types/crew.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Crew {
  id: string;
  companyId: string;
  name: string;
  supervisorUserId: string;
  isActive: boolean;
  createdAt: Date;
  updatedAt: Date;
}