# lib/api-zod/src/generated/types/crewPatch.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CrewPatch {
  /**
   * @minLength 1
   * @maxLength 120
   */
  name?: string;
  supervisorUserId?: string;
  isActive?: boolean;
}