# lib/api-zod/src/generated/types/crewWithSupervisor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Crew } from "./crew";

export type CrewWithSupervisor = Crew & {
  supervisorName?: string | null;
  supervisorEmail?: string | null;
};