# lib/api-zod/src/generated/types/eligibleSupervisor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from "./role";

export interface EligibleSupervisor {
  id: string;
  name: string;
  email?: string | null;
  phone?: string | null;
  role: Role;
}