# lib/api-zod/src/generated/types/mobileSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MobileUser } from "./mobileUser";

export interface MobileSession {
  /** Opaque bearer token to be sent as `Authorization: Bearer <token>` on subsequent /api/m/* requests. */
  token: string;
  expiresAt: Date;
  user: MobileUser;
}