# lib/api-zod/src/generated/types/mobileUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MobileUserLanguage } from "./mobileUserLanguage";
import type { Role } from "./role";

export interface MobileUser {
  id: string;
  name: string;
  email?: string | null;
  phone?: string | null;
  language?: MobileUserLanguage;
  activeCompanyId: string;
  activeRole: Role;
  isSuperAdminBool?: boolean;
}