# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  admin: "admin",
  office: "office",
  field_manager: "field_manager",
  chemical_manager: "chemical_manager",
  field: "field",
  irrigation_manager: "irrigation_manager",
  shop_manager: "shop_manager",
  mapping: "mapping",
  landscape_supervisor: "landscape_supervisor",
  crew_supervisor: "crew_supervisor",
} as const;